feat(backup): restore backup struct
